data program/data_storage/requirements.txt
cassandra-driver==3.28.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
kafka-python==2.0.2